Fix type conversion errors in WeeklyPlanService and update weekly plan models
- Fix CS0266 errors: Add explicit int casts for long to int conversions in WeeklyPlanService
- Update WeeklyPlanService to handle type conversions between model (long) and DTO (int) IDs
- Add DailyProgress model and repository support
- Update WeeklyPlan and WeeklyPlanTask models with new properties
- Add WeeklyPlan DTOs and update related services
- Update TaskProgress and related models
- Add new SQL scripts and documentation
- Remove obsolete SEED_SALES_ACTIVITIES.sql file

+++ SoitMed/Controllers/TaskProgressController.cs
@@ -172,7 +172,7 @@ namespace SoitMed.Controllers
         /// </summary>
         [HttpGet("task/{taskId}")]
         [Authorize(Roles = "SalesMan,SalesManager,SuperAdmin")]
-        public async Task<IActionResult> GetProgressByTask(long taskId)
+        public async Task<IActionResult> GetProgressByTask(int taskId)
         {
             try
             {
