+++ SoitMed/Repositories/IWeeklyPlanRepository.cs
@@ -5,9 +5,9 @@ namespace SoitMed.Repositories
 {
     public interface IWeeklyPlanRepository
     {
-        Task<WeeklyPlan?> GetByIdAsync(int id, CancellationToken cancellationToken = default);
-        Task<WeeklyPlan?> GetByIdWithDetailsAsync(int id, CancellationToken cancellationToken = default);
-        Task<WeeklyPlan?> GetByIdAndEmployeeIdAsync(int id, string employeeId, CancellationToken cancellationToken = default);
+        Task<WeeklyPlan?> GetByIdAsync(long id, CancellationToken cancellationToken = default);
+        Task<WeeklyPlan?> GetByIdWithDetailsAsync(long id, CancellationToken cancellationToken = default);
+        Task<WeeklyPlan?> GetByIdAndEmployeeIdAsync(long id, string employeeId, CancellationToken cancellationToken = default);
         Task<IEnumerable<WeeklyPlan>> GetAllAsync(CancellationToken cancellationToken = default);
         Task<IEnumerable<WeeklyPlan>> GetByEmployeeIdAsync(string employeeId, CancellationToken cancellationToken = default);
         Task<(IEnumerable<WeeklyPlan> Plans, int TotalCount)> GetPaginatedAsync(
@@ -17,10 +17,10 @@ namespace SoitMed.Repositories
             CancellationToken cancellationToken = default);
         Task<WeeklyPlan> CreateAsync(WeeklyPlan weeklyPlan, CancellationToken cancellationToken = default);
         Task<WeeklyPlan> UpdateAsync(WeeklyPlan weeklyPlan, CancellationToken cancellationToken = default);
-        Task<bool> DeleteAsync(int id, CancellationToken cancellationToken = default);
-        Task<bool> ExistsAsync(int id, CancellationToken cancellationToken = default);
-        Task<bool> ExistsForEmployeeAsync(int id, string employeeId, CancellationToken cancellationToken = default);
-        Task<bool> HasPlanForWeekAsync(string employeeId, DateOnly weekStartDate, int? excludePlanId = null, CancellationToken cancellationToken = default);
+        Task<bool> DeleteAsync(long id, CancellationToken cancellationToken = default);
+        Task<bool> ExistsAsync(long id, CancellationToken cancellationToken = default);
+        Task<bool> ExistsForEmployeeAsync(long id, string employeeId, CancellationToken cancellationToken = default);
+        Task<bool> HasPlanForWeekAsync(string employeeId, DateOnly weekStartDate, long? excludePlanId = null, CancellationToken cancellationToken = default);
     }
 }
